+++ lib/src/models/response/pixQrCode/abcatepay_pixqrcode_status_response.dart
@@ -1,5 +1,6 @@
 import 'package:abacatepay/src/models/constants/abacatepay_constants.dart';
 import 'package:abacatepay/src/models/enums/payment_status_enum.dart';
+import 'package:abacatepay/src/utils/get_enum_by_apikey.dart';
 
 final class AbacatePayPixQrCodeStatusResponse {
   final PaymentStatus status;
@@ -13,7 +14,11 @@ final class AbacatePayPixQrCodeStatusResponse {
   factory AbacatePayPixQrCodeStatusResponse.fromData(
     Map<String, dynamic> data,
   ) => AbacatePayPixQrCodeStatusResponse(
-    status: data[AbacatePayConstants.status],
+    status:
+        getEnumByApiKey<PaymentStatus>(
+          PaymentStatus.values,
+          data[AbacatePayConstants.status],
+        )!,
     expiresAt: DateTime.parse(data[AbacatePayConstants.expiresAt]),
   );
 }