+++ CHANGELOG.md
@@ -1,3 +1,14 @@
 ## 1.0.0
 
-- Initial version.
+Fully integrated with all AbacatePay API usecases
+
+- Customers usecases
+  - Create new customers
+  - List customers
+- Billing usecases
+  - Create new billing
+  - List billings
+- PixQrCode usecases
+  - Create QrCode PIX
+  - Simulate payment
+  - Check status